tag_tree_node_create refactoring
tag_tree_node_create now takes its field parameters as arguments rather than
having the caller set them.  We still hold the lock for the caller to
unlock in order for the data itself to be set by the time other threads can
read the tag.

This is work in anticipation of changing the public interface such that
the tag tree lock can be held during tag allocation and initialisation in
plc_tag_create().