Per-tag type tracking
This patch introduces an associated type with every tag and uses
type information for simple checks like offset calculation and
out-of-bounds checks.

This patch also introduces loading types from an external file,
called `tags.inc`.  Currently I need to think through the
mechanism for arbitrary types, but James and I have discussed
"flattening" out complex types, which this is certainly sufficient
for.

This required applying a double-bladed lightsabre to this code,
so the end result is not as tidy.  There will be other patches
to decouple tag IDs from tags which will further mess things up;
once those are in I will come back and tidy stuff up.